requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
backoff==2.2.1
bcrypt==4.2.1
blinker>=1.9
build==1.2.2.post1
cachetools==5.5.0
certifi>=2024.8
cffi>=1.17.1
charset-normalizer>=3.4.0
chroma-hnswlib==0.7.6
chromadb==0.5.20
click>=8.1.7
cmake>=3.30.0
colorama>=0.4.6
coloredlogs>=15.0.1
colormath>=3.0.0
contourpy>=1.3.1
cycler>=0.12.1
Deprecated>=1.2.15
dlib>=19.24.6
durationpy>=0.9
fastapi>=0.115.5
filelock>=3.16.1
Flask>=3.1.0
flatbuffers>=24.3.0
fonttools>=4.54.1
fsspec>=2024.10.0
google-auth>=2.36.0
googleapis-common-protos>=1.66.0
grpcio>=1.68.0
h11>=0.14.0
httpcore>=1.0.7
httptools>=0.6.4
httpx>=0.27.2
huggingface-hub>=0.26.2
humanfriendly>=10.0
idna>=3.10
importlib_metadata>=8.5.0
importlib_resources>=6.4.5
itsdangerous>=2.2.0
Jinja2>=3.1.4
joblib>=1.4.2
kiwisolver>=1.4.7
kubernetes==31.0.0
markdown-it-py>=3.0.0
MarkupSafe>=3.0.2
matplotlib>=3.9.2
mdurl>=0.1.2
mmh3>=5.0.1
monotonic>=1.6
mpmath>=1.3.0
mysql-connector-python>=9.1.0
networkx>=3.4.2
numpy>=1.21.0
oauthlib>=3.2.2
onnxruntime>=1.20.1
opencv-python>=4.10.0.84
opentelemetry-api>=1.28.2
opentelemetry-exporter-otlp-proto-common>=1.28.2
opentelemetry-exporter-otlp-proto-grpc>=1.28.2
opentelemetry-instrumentation>=0.49b2
opentelemetry-instrumentation-asgi>=0.49b2
opentelemetry-instrumentation-fastapi>=0.49b2
opentelemetry-proto>=1.28.2
opentelemetry-sdk>=1.28.2
opentelemetry-semantic-conventions>=0.49b2
opentelemetry-util-http>=0.49b2
orjson>=3.10.12
overrides>=7.7.0
packaging>=24.2
pillow>=11.0.0
posthog>=3.7.2
protobuf>=3.20.0
pyasn1>=0.6.1
pyasn1_modules>=0.4.1
pydantic>=2.10.1
pydantic_core>=2.27.1
Pygments>=2.18.0
pyparsing>=3.2.0
PyPika>=0.48.9
pyproject_hooks>=1.2.0
pyreadline3>=3.5.4
Pyrebase4==4.8.0
python-dateutil>=2.9.0.post0
python-dotenv>=1.0.1
PyYAML>=6.0.2
requests>=2.32.3
requests-oauthlib>=2.0.0
rich>=13.9.4
rsa>=4.9
scikit-learn>=1.0
scipy>=1.8.0
shellingham>=1.5.4
six>=1.16.0
sniffio>=1.3.1
starlette>=0.41.3
sympy>=1.13.3
tenacity>=9.0.0